requirements.txt
ahk==0.12.0
altgraph==0.17
future==0.18.2
Jinja2==3.0.1
MarkupSafe==2.0.1
MouseInfo==0.1.3
mss==6.1.0
Nuitka==0.6.16.3
numpy==1.21.1
opencv-python==4.5.3.56
pefile==2021.5.24
Pillow==8.3.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
pytesseract==0.3.8
PyTweening==1.0.3
pywin32-ctypes==0.2.0